@@ -1,6 +1,6 @@
 module github.com/jkaninda/mysql-bkup
 
-go 1.22.5
+go 1.23.2
 
 require github.com/spf13/pflag v1.0.5 // indirect
 
