@@ -26,6 +26,7 @@ package internal
 
 import (
 	"fmt"
+	"github.com/jkaninda/mysql-bkup/pkg/logger"
 	"github.com/jkaninda/mysql-bkup/utils"
 	"github.com/spf13/cobra"
 	"os"
@@ -104,7 +105,7 @@ type AWSConfig struct {
 }
 
 func initDbConfig(cmd *cobra.Command) *dbConfig {
-	//Set env
+	// Set env
 	utils.GetEnv(cmd, "dbname", "DB_NAME")
 	dConf := dbConfig{}
 	dConf.dbHost = os.Getenv("DB_HOST")
@@ -115,8 +116,8 @@ func initDbConfig(cmd *cobra.Command) *dbConfig {
 
 	err := utils.CheckEnvVars(dbHVars)
 	if err != nil {
-		utils.Error("Please make sure all required environment variables for database are set")
-		utils.Fatal("Error checking environment variables: %s", err)
+		logger.Error("Please make sure all required environment variables for database are set")
+		logger.Fatal("Error checking environment variables: %s", err)
 	}
 	return &dConf
 }
@@ -149,7 +150,7 @@ func loadSSHConfig() (*SSHConfig, error) {
 	}, nil
 }
 func loadFtpConfig() *FTPConfig {
-	//Initialize data configs
+	// Initialize data configs
 	fConfig := FTPConfig{}
 	fConfig.host = utils.GetEnvVariable("FTP_HOST", "FTP_HOST_NAME")
 	fConfig.user = os.Getenv("FTP_USER")
@@ -158,13 +159,13 @@ func loadFtpConfig() *FTPConfig {
 	fConfig.remotePath = os.Getenv("REMOTE_PATH")
 	err := utils.CheckEnvVars(ftpVars)
 	if err != nil {
-		utils.Error("Please make sure all required environment variables for FTP are set")
-		utils.Fatal("Error missing environment variables: %s", err)
+		logger.Error("Please make sure all required environment variables for FTP are set")
+		logger.Fatal("Error missing environment variables: %s", err)
 	}
 	return &fConfig
 }
 func initAWSConfig() *AWSConfig {
-	//Initialize AWS configs
+	// Initialize AWS configs
 	aConfig := AWSConfig{}
 	aConfig.endpoint = utils.GetEnvVariable("AWS_S3_ENDPOINT", "S3_ENDPOINT")
 	aConfig.accessKey = utils.GetEnvVariable("AWS_ACCESS_KEY", "ACCESS_KEY")
@@ -185,8 +186,8 @@ func initAWSConfig() *AWSConfig {
 	aConfig.forcePathStyle = forcePathStyle
 	err = utils.CheckEnvVars(awsVars)
 	if err != nil {
-		utils.Error("Please make sure all required environment variables for AWS S3 are set")
-		utils.Fatal("Error checking environment variables: %s", err)
+		logger.Error("Please make sure all required environment variables for AWS S3 are set")
+		logger.Fatal("Error checking environment variables: %s", err)
 	}
 	return &aConfig
 }
@@ -194,7 +195,7 @@ func initBackupConfig(cmd *cobra.Command) *BackupConfig {
 	utils.SetEnv("STORAGE_PATH", storagePath)
 	utils.GetEnv(cmd, "cron-expression", "BACKUP_CRON_EXPRESSION")
 	utils.GetEnv(cmd, "path", "REMOTE_PATH")
-	//Get flag value and set env
+	// Get flag value and set env
 	remotePath := utils.GetEnvVariable("REMOTE_PATH", "SSH_REMOTE_PATH")
 	storage = utils.GetEnv(cmd, "storage", "STORAGE")
 	prune := false
@@ -216,7 +217,7 @@ func initBackupConfig(cmd *cobra.Command) *BackupConfig {
 		encryption = true
 		usingKey = false
 	}
-	//Initialize backup configs
+	// Initialize backup configs
 	config := BackupConfig{}
 	config.backupRetention = backupRetention
 	config.disableCompression = disableCompression
@@ -246,7 +247,7 @@ func initRestoreConfig(cmd *cobra.Command) *RestoreConfig {
 	utils.SetEnv("STORAGE_PATH", storagePath)
 	utils.GetEnv(cmd, "path", "REMOTE_PATH")
 
-	//Get flag value and set env
+	// Get flag value and set env
 	s3Path := utils.GetEnv(cmd, "path", "AWS_S3_PATH")
 	remotePath := utils.GetEnvVariable("REMOTE_PATH", "SSH_REMOTE_PATH")
 	storage = utils.GetEnv(cmd, "storage", "STORAGE")
@@ -260,7 +261,7 @@ func initRestoreConfig(cmd *cobra.Command) *RestoreConfig {
 		usingKey = false
 	}
 
-	//Initialize restore configs
+	// Initialize restore configs
 	rConfig := RestoreConfig{}
 	rConfig.s3Path = s3Path
 	rConfig.remotePath = remotePath
@@ -276,15 +277,15 @@ func initRestoreConfig(cmd *cobra.Command) *RestoreConfig {
 func initTargetDbConfig() *targetDbConfig {
 	tdbConfig := targetDbConfig{}
 	tdbConfig.targetDbHost = os.Getenv("TARGET_DB_HOST")
-	tdbConfig.targetDbPort = utils.EnvWithDefault("TARGET_DB_PORT", "5432")
+	tdbConfig.targetDbPort = utils.EnvWithDefault("TARGET_DB_PORT", "3306")
 	tdbConfig.targetDbName = os.Getenv("TARGET_DB_NAME")
 	tdbConfig.targetDbUserName = os.Getenv("TARGET_DB_USERNAME")
 	tdbConfig.targetDbPassword = os.Getenv("TARGET_DB_PASSWORD")
 
 	err := utils.CheckEnvVars(tdbRVars)
 	if err != nil {
-		utils.Error("Please make sure all required environment variables for the target database are set")
-		utils.Fatal("Error checking target database environment variables: %s", err)
+		logger.Error("Please make sure all required environment variables for the target database are set")
+		logger.Fatal("Error checking target database environment variables: %s", err)
 	}
 	return &tdbConfig
 }
