@@ -1,8 +1,9 @@
-// Package cmd /
+package logger
+
 /*
 MIT License
 
-Copyright (c) 2023 Jonas Kaninda
+# Copyright (c) 2023 Jonas Kaninda
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -22,33 +23,4 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-package cmd
-
-import (
-	"github.com/jkaninda/mysql-bkup/internal"
-	"github.com/jkaninda/mysql-bkup/pkg/logger"
-	"github.com/jkaninda/mysql-bkup/utils"
-	"github.com/spf13/cobra"
-)
-
-var BackupCmd = &cobra.Command{
-	Use:     "backup ",
-	Short:   "Backup database operation",
-	Example: utils.BackupExample,
-	Run: func(cmd *cobra.Command, args []string) {
-		if len(args) == 0 {
-			internal.StartBackup(cmd)
-		} else {
-			logger.Fatal(`"backup" accepts no argument %q`, args)
-		}
-	},
-}
-
-func init() {
-	//Backup
-	BackupCmd.PersistentFlags().StringP("storage", "s", "local", "Define storage: local, s3, ssh, ftp")
-	BackupCmd.PersistentFlags().StringP("path", "P", "", "AWS S3 path without file name. eg: /custom_path or ssh remote path `/home/foo/backup`")
-	BackupCmd.PersistentFlags().StringP("cron-expression", "", "", "Backup cron expression")
-	BackupCmd.PersistentFlags().BoolP("disable-compression", "", false, "Disable backup compression")
-
-}
+const traceLog = "trace"
